end7

+++ controller/Categories.js
@@ -2,20 +2,21 @@
 
 import Categories from "../model/Categories.js"
 import cloudinary from 'cloudinary';
+import User from "../model/User.js";
 
 
 export const getAll = async (req, res) => {
-   
-   
+
+
    try {
       let currentPage = Number(req.query.page) || 1;
       const totalElements = await Categories.countDocuments();
-      const elementPerPage = currentPage === -1 ? totalElements : (Number(req.query.PerPage ) || 4);
+      const elementPerPage = currentPage === -1 ? totalElements : (Number(req.query.PerPage) || 4);
       const totalPage = Math.ceil(totalElements / elementPerPage);
 
-      
-      if(currentPage < 0)
-         currentPage = 1; 
+
+      if (currentPage < 0)
+         currentPage = 1;
 
 
       const categories = await Categories.find().skip((currentPage - 1) * elementPerPage).limit(elementPerPage);
@@ -42,6 +43,13 @@ export const get = async (req, res) => {
 }
 
 export const create = async (req, res) => {
+
+   const { email } = req.user;
+   const user = await User.findOne({ email });
+
+   if (user.isAdmin === false) {
+      return res.status(400).json({ success: false, msg: 'Only admin can create category' });
+   }
    try {
       const images = req.body.images;
 
@@ -77,6 +85,12 @@ export const create = async (req, res) => {
 
 export const deletes = async (req, res) => {
 
+   const { email } = req.user;
+   const user = await User.findOne({ email });
+
+   if (user.isAdmin === false) {
+      return res.status(400).json({ success: false, msg: 'Only admin can delete category' });
+   }
    try {
       const id = req.params.id;
       const category = await Categories.findByIdAndDelete(id);
@@ -91,6 +105,12 @@ export const deletes = async (req, res) => {
 
 export const update = async (req, res) => {
 
+      const {email} = req.user;
+   const user = await User.findOne({email});
+
+   if(user.isAdmin === false){
+      return res.status(400).json({success : false,msg : 'Only admin can edit category'});
+   }
    try {
       const images = req.body.images;
